# docs/swagger.yaml
basePath: /
definitions:
...
  entities.ListCreate:
    properties:
      title:
        type: string
    required:
    - title
    type: object
  entities.ListGet:
    properties:
      id:
        type: integer
      title:
        type: string
      userId:
        type: integer
    type: object
  entities.ListGetWithItems:
    properties:
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/entities.ListItemGet'
        type: array
      title:
        type: string
      userId:
        type: integer
    type: object
  entities.ListItemCreate:
    properties:
      bookId:
        type: integer
      status:
        type: string
    required:
    - bookId
    - status
    type: object
  entities.ListItemGet:
    properties:
      bookId:
        type: integer
      status:
        type: string
    type: object
  entities.ListItemUpdate:
    properties:
      status:
        type: string
    type: object
  entities.ListUpdate:
    properties:
      title:
        type: string
    type: object
  entities.Lists:
    properties:
      data:
        items:
          $ref: '#/definitions/entities.ListGet'
        type: array
    type: object
...
  handler.ResponseStruct:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: RESTful API service for library management system
  title: Library app API
  version: "1.0"
paths:
...
  /api/lists:
    get:
      description: Get all the users' book lists in the system
      produces:
      - application/json
      responses:
        "200":
          description: users' book lists data
          schema:
            $ref: '#/definitions/entities.Lists'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ResponseStruct'
      security:
      - ApiKeyAuth: []
      summary: Get user's book lists
      tags:
      - Lists
    post:
      consumes:
      - application/json
      description: Create new user's book list in the system
      parameters:
      - description: new user's book list info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entities.ListCreate'
      produces:
      - application/json
      responses:
        "200":
          description: new user's book list id
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ResponseStruct'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ResponseStruct'
      security:
      - ApiKeyAuth: []
      summary: Create user's book list
      tags:
      - Lists
  /api/lists/:list:
    delete:
      description: Delete user's book list by id in the system
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ResponseStruct'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ResponseStruct'
      security:
      - ApiKeyAuth: []
      summary: Delete user's book list
      tags:
      - Lists
    get:
      description: Get the user's book list by id in the system
      produces:
      - application/json
      responses:
        "200":
          description: user's book list data
          schema:
            $ref: '#/definitions/entities.ListGetWithItems'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ResponseStruct'
      security:
      - ApiKeyAuth: []
      summary: Get user's book list
      tags:
      - Lists
    put:
      consumes:
      - application/json
      description: Update user's book list data by id in the system
      parameters:
      - description: new user's book list data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entities.ListUpdate'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ResponseStruct'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ResponseStruct'
      security:
      - ApiKeyAuth: []
      summary: Update user's book list
      tags:
      - Lists
  /api/lists/:list/items:
    post:
      consumes:
      - application/json
      description: Create new list item in the system
      parameters:
      - description: new list item info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entities.ListItemCreate'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ResponseStruct'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ResponseStruct'
      security:
      - ApiKeyAuth: []
      summary: Create list item
      tags:
      - Items
  /api/lists/:list/items/:book:
    delete:
      description: Delete list item by id in the system
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ResponseStruct'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ResponseStruct'
      security:
      - ApiKeyAuth: []
      summary: Delete list item
      tags:
      - Items
    put:
      consumes:
      - application/json
      description: Update list item data by id in the system
      parameters:
      - description: new genre data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entities.ListItemUpdate'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ResponseStruct'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ResponseStruct'
      security:
      - ApiKeyAuth: []
      summary: Update list item
      tags:
      - Items
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"